feat(node/engine): decouple engine and runtime actor
